script to transform a version of hulu data
really, it's just a csv with fields named "Episode Name", "Series Name"
(which can be "N/A" for movies), Season (which should also be "N/A" for
movies), and "Last Played At". This format is somewhat easy to get from
the data from Hulu's CCPA report with a bit of work.